Keylogger/requirements.txt
pynput==1.7.6
python-dateutil==2.8.2
PyQt5==5.15.9
pywin32==306
pyinstaller==6.3.0
Pillow==10.2.0
pyperclip==1.8.2
cryptography==42.0.2 